% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rspatialvis-package.R
\docType{package}
\name{rspatialvis-package}
\alias{rspatialvis}
\alias{rspatialvis-package}
\title{rspatialvis: Interactive Visualization for Spatial Datasets}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Christopher Ly \email{chrisly92@gmail.com}

}
\keyword{internal}
